+++ src/main/java/hoggaster/rules/indicators/SimpleValueIndicator.java
@@ -3,21 +3,23 @@ package hoggaster.rules.indicators;
 import com.google.common.base.Preconditions;
 import hoggaster.robot.RobotExecutionContext;
 
+import java.math.BigDecimal;
+
 /**
  * Simple fixed single value indicator, like '2.0'. Can be used
  * to trigger actions base on a fixed price like fixed stop loss.
  */
 public class SimpleValueIndicator implements Indicator {
 
-    public final Double value;
+    public final BigDecimal value;
 
-    public SimpleValueIndicator(Double value) {
+    public SimpleValueIndicator(BigDecimal value) {
         Preconditions.checkNotNull(value);
         this.value = value;
     }
 
     @Override
-    public Double value(RobotExecutionContext ctx) {
+    public BigDecimal value(RobotExecutionContext ctx) {
         return value;
     }
